Working REST API withouth authentication, Automapper and Documentation
REST API is now working without authentication. Services were tested via Postman, Swagger and HttpClient.
@TODO: User and Service authentication and authorization

Added automapper to translate between client models and data models.
@TODO: Add a new layer between business and data models

Created APIHandler that handles all communication between internal applications and external APIs using HttpClient.
Added TogglerAPI files including a settings file that will handle services calls to the API

Created a Common project where lives a generic SettingsReader and an ApplicationInformation that gather assembly information

Did a bit of code cleanup and documentation.

+++ Farfetch.Repositories/MongoDB/MongoRepository.cs
@@ -2,17 +2,28 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Linq.Expressions;
-using Farfetch.DB.MongoDB;
+using Farfetch.DB;
 using Farfetch.Models;
 using MongoDB.Driver;
 
 namespace Farfetch.Repositories.MongoDB
 {
+    /// <inheritdoc />
+    /// <summary>
+    /// Defines a MongoDatabase
+    /// </summary>
     public class MongoRepository<T> : IRepository<T> where T: DbT
     {
+        /// <summary>
+        /// The collection matching T
+        /// </summary>
         private readonly IMongoCollection<T> _collection;
 
-        public MongoRepository(MongoDatabase<IMongoDatabase> database)
+        /// <summary>
+        /// Default construtor that initializes the database
+        /// </summary>
+        /// <param name="database">The database information</param>
+        public MongoRepository(IDatabase<IMongoDatabase> database)
         {
             if (database == null) throw new ArgumentNullException(nameof(database));
             IMongoDatabase db = database.GetDatabase();
@@ -24,48 +35,56 @@ namespace Farfetch.Repositories.MongoDB
         /// <inheritdoc />
         public List<T> GetAll()
         {
-            return _collection?.AsQueryable()?.ToList();
+            if (_collection == null) throw new NullReferenceException("Failed to initialize the collection");
+            return _collection.AsQueryable()?.ToList();
         }
 
         /// <inheritdoc />
         public List<T> GetAllFiltered(Expression<Func<T, bool>> expression)
         {
-            return _collection?.Find(expression).ToList();
+            if (_collection == null) throw new NullReferenceException("Failed to initialize the collection");
+            return _collection.Find(expression).ToList();
         }
 
         /// <inheritdoc />
         public T GetSingle(Expression<Func<T, bool>> expression)
         {
-            return _collection.Find(expression).ToList().FirstOrDefault();
+            if (_collection == null) throw new NullReferenceException("Failed to initialize the collection");
+            return _collection.Find(expression)?.ToList()?.FirstOrDefault();
         }
 
         /// <inheritdoc />
         public void Insert(T value)
         {
+            if (_collection == null) throw new NullReferenceException("Failed to initialize the collection");
             _collection.InsertOne(value);
         }
 
         /// <inheritdoc />
         public void Update(T value)
         {
+            if (_collection == null) throw new NullReferenceException("Failed to initialize the collection");
             _collection.ReplaceOne(x => x.Id == value.Id, value);
         }
 
         /// <inheritdoc />
         public void Update(Expression<Func<T, bool>> expression, T value)
         {
+            if (_collection == null) throw new NullReferenceException("Failed to initialize the collection");
             _collection.ReplaceOne(expression, value);
         }
 
         /// <inheritdoc />
         public void Delete(Expression<Func<T, bool>> expression)
         {
+            if (_collection == null) throw new NullReferenceException("Failed to initialize the collection");
             _collection.DeleteOne(expression);
         }
 
         /// <inheritdoc />
         public void DeleteMany(Expression<Func<T, bool>> expression)
         {
+            if (_collection == null) throw new NullReferenceException("Failed to initialize the collection");
             _collection.DeleteMany(expression);
         }
     }
